+++ src/app/dashboard/[dashboardid]/edit/page.tsx
@@ -1,7 +1,10 @@
 import Image from 'next/image';
-import EditDashboard from '@/src/components/dashboardEdit/EditDashboard';
-import MemberList from '@/src/components/dashboardEdit/MemberList';
-import InvitationList from '@/src/components/dashboardEdit/InvitationList';
+import Link from 'next/link';
+
+// eslint-disable-next-line sort-imports
+import EditDashboard from '@/src/components/edit/EditDashboard';
+import InvitationList from '@/src/components/edit/InvitationList';
+import MemberList from '@/src/components/edit/MemberList';
 
 export default function DashBoardEditPage() {
   return (
@@ -10,12 +13,13 @@ export default function DashBoardEditPage() {
         헤더
       </header>
       <div className="flex flex-1">
-        <aside className="w-[67px] bg-gray-100 md:block md:w-[160px] xl:w-[300px]">
+        <aside className="w-[67px] bg-gray-200 md:block md:w-[160px] xl:w-[300px]">
           사이드바
         </aside>
 
-        <main className="flex-1 overflow-y-auto p-4">
-          <div className="mb-4 flex items-center gap-2">
+        <main className="flex-1 overflow-y-auto bg-gray-50 p-4">
+          {/* 경로 임시로 지정해둔 것 */}
+          <Link href="/" className="mb-4 flex items-center gap-2">
             <Image
               src="/icons/arrow_left.png"
               alt="돌아가기"
@@ -23,14 +27,14 @@ export default function DashBoardEditPage() {
               height={16}
             />
             <span className="font-lg-16px-medium">돌아가기</span>
-          </div>
+          </Link>
           <div className="space-y-6">
             <EditDashboard />
             <MemberList />
             <InvitationList />
             <button
               type="button"
-              className="flex h-[52px] w-full max-w-md items-center justify-center rounded-lg border border-solid border-gray-200 font-lg-16px-medium md:w-[320px]"
+              className="flex h-[52px] w-full max-w-md items-center justify-center rounded-lg border border-solid border-gray-200 shadow font-lg-16px-medium md:w-[320px]"
             >
               대시보드 삭제하기
             </button>
